app/requirements.txt
Django>=3.1.3,<3.2.0
djangorestframework>=3.12.2,<3.13.0
flake8>=3.8.4,<3.9.0
psycopg2>=2.8.6,<2.9.0
django-cors-headers>=3.5.0,<3.6.0
django-debug-toolbar>=3.1.1,<3.2.0
drf-yasg>=1.20.0,<1.21.0
martor>=1.5.7,<1.6.0
Markdown>=3.3.3,<3.4.0
boto3>=1.16.23,<1.17.0
django-storages>=1.10.1,<1.11.0
djangorestframework-simplejwt>=4.6.0
pytz~=2020.4
requests~=2.25.0
django-3-jet==1.0.8
django-redis>=4.12.1,<4.13.0
django-redisboard>=4.1.1,<4.2.0
djoser>=2.0.5,<2.1.0
requests_futures==1.0.0
django-silk==4.1.0